Implement Data and Repo Layers
Implemented Data Layers and Model definitions to develop a rudementary repository layer which uses the repository design pattern to partition business logic from data retrieval.

All funcs in the UserRepository have passed testing. This is just a basic implementation so progress can happen on the fun stuff, the service layer. 💀